Various changes to make the DatabaseExplorer build on wx2.9, including changing an event table: you can't use XRCIDs in an EVT_*_RANGE event table entry; it was always likely to fail, and in wx2.9 it asserts
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@4854 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ DatabaseExplorer/databaseexplorer.cpp
@@ -201,7 +201,7 @@ void DatabaseExplorer::UnPlug()
 	//TODO:: perform the unplug action for this plugin
 	
 	
-	int index = m_mgr->GetWorkspacePaneNotebook()->GetPageIndex(m_dbViewerPanel);
+	size_t index = m_mgr->GetWorkspacePaneNotebook()->GetPageIndex(m_dbViewerPanel);
 	if (index != Notebook::npos) {
 		m_mgr->GetWorkspacePaneNotebook()->RemovePage(index, false);
 	}
